requirements.txt
streamlit
transformers
torch
pypdf2
python-docx
python-pptx